requirements.txt
discord.py
requests
pickleDB
apscheduler
